+++ catboost/libs/helpers/query_info_helper.h
@@ -8,4 +8,4 @@ void UpdateQueriesInfo(const TVector<ui32>& queriesId, int begin, int end, TVect
 
 TVector<int> GetQueryIndicesForDocs(const TVector<TQueryInfo>& queriesInfo, int learnSampleCount);
 
-void UpdateQueriesPairs(const TVector<TPair>& pairs, int begin, int end, const TVector<int>& invertedPermutation, TVector<TQueryInfo>* queryInfo);
+void UpdateQueriesPairs(const TVector<TPair>& pairs, int begin, int end, const TVector<size_t>& invertedPermutation, TVector<TQueryInfo>* queryInfo);
